Use ast to parse the tree
This has the benefit that we ensure it only checks actual function
calls.

As far as I can tell this means we don't need the filename at all
anymore.

The reason one of the tests has changed the column offset from 9 to
0 is that the tree measures offsets in a different way to the string
matching.

I've followed the pattern from flake8-pytest quite closely in this
commit:
https://github.com/vikingco/flake8-pytest/blob/c6861968d7ee8fd9680343f6bae5fb/flake8_pytest.py